Updated file relocation changes in manifest and package paths
